+++ packages/fluent_environment/fluent_environment/lib/src/widgets/environment_banner.dart
@@ -2,33 +2,52 @@ import 'package:fluent_environment_api/fluent_environment_api.dart';
 import 'package:fluent_sdk/fluent_sdk.dart';
 import 'package:flutter/material.dart';
 
-/// Widget that use the current environment to display
-/// a banner with the environment name and color at the bottom of the screen
+/// A widget that displays an environment banner if
+/// the current environment is not production.
+///
+/// This widget wraps its [child] and, if the detected or provided [environment]
+/// is not production, it overlays a [Banner]
+/// with the environment's name and color.
+/// The banner is typically displayed at the bottom-end of the widget.
 class EnvironmentBanner extends StatelessWidget {
+  /// Creates an [EnvironmentBanner].
+  ///
+  /// The [child] widget is always required.
+  /// The [environment] can be optionally provided; if null, it defaults to
+  /// the environment retrieved from `Fluent.get<EnvironmentApi>()`.
   const EnvironmentBanner({
     required this.child,
+    this.environment,
     super.key,
   });
 
-  /// Widget that will be inside of the banner
+  /// The widget below this widget in the tree.
+  ///
+  /// This is the content that the environment banner will overlay.
   final Widget child;
+
+  /// The environment to display.
+  ///
+  /// If `null`, the environment will be retrieved from the `Fluent` instance
+  /// using `Fluent.get<EnvironmentApi>().environment`.
+  final Environment? environment;
 
   @override
   Widget build(BuildContext context) {
-    final environment = Fluent.get<EnvironmentApi>().environment;
+    final env = environment ?? Fluent.get<EnvironmentApi>().environment;
 
-    if (environment.type == EnvironemntType.prod) {
+    if (env.isProduction) {
       return child;
     }
 
     return Directionality(
       textDirection: TextDirection.ltr,
       child: Banner(
-        color: environment.color,
-        message: environment.name,
+        color: env.color,
+        message: env.name,
         location: BannerLocation.bottomEnd,
         textStyle: const TextStyle(
-          fontSize: 12.0 * 0.85,
+          fontSize: 10.2,
           fontWeight: FontWeight.w900,
           height: 1,
         ),