+++ packages/fluent_logger/fluent_logger/lib/flutter_fluent_logger.dart
@@ -1,4 +1,5 @@
 export 'package:fluent_logger_api/fluent_logger_api.dart';
 export 'package:fluent_sdk/fluent_sdk.dart';
 
+export 'src/logger_config.dart';
 export 'src/logger_module.dart';