+++ packages/fluent_logger/fluent_logger/lib/src/logger_config.dart
@@ -1,0 +1,18 @@
+class LoggerConfig {
+  /// Creates a configuration for the Logger module.
+  ///
+  /// [enableLog]: If true, logs will be printed to the console.
+  /// usually controlled by `kDebugMode` in Flutter apps.
+  /// [globalLogName]: The tag used for the logs (defaults to 'App').
+  LoggerConfig({
+    this.enableLog = false,
+    this.globalLogName = 'App',
+  });
+
+  /// If true, logs will be printed to the console.
+  /// usually controlled by `kDebugMode` in Flutter apps.
+  final bool enableLog;
+
+  /// The tag used for the logs (defaults to 'App').
+  final String globalLogName;
+}