+++ packages/fluent_logger/fluent_logger/test/src/logger_module_test.dart
@@ -1,19 +1,32 @@
 import 'package:fluent_logger_api/fluent_logger_api.dart';
 import 'package:fluent_sdk/fluent_sdk.dart';
 import 'package:flutter_fluent_logger/src/api/logger_api_impl.dart';
+import 'package:flutter_fluent_logger/src/logger_config.dart';
 import 'package:flutter_fluent_logger/src/logger_module.dart';
 import 'package:loggy/loggy.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('verify logger module', () async {
+  test('verify logger module registration', () async {
+    // Arrange
+    final config = LoggerConfig(enableLog: true);
+
+    // Act
+    await Fluent.build([LoggerModule(config: config)]);
+    addTearDown(Fluent.reset);
+
+    // Assert
+    expect(Fluent.get<Loggy>(), isA<Loggy>());
+    expect(Fluent.get<LoggerApi>(), isA<LoggerApiImpl>());
+  });
+
+  test('verify logger module registration without config', () async {
+    // Act
     await Fluent.build([LoggerModule()]);
     addTearDown(Fluent.reset);
 
-    final loggy = Fluent.get<Loggy>();
-    final loggerApi = Fluent.get<LoggerApi>();
-
-    expect(loggy, isA<Loggy>());
-    expect(loggerApi, isA<LoggerApiImpl>());
+    // Assert
+    expect(Fluent.get<Loggy>(), isA<Loggy>());
+    expect(Fluent.get<LoggerApi>(), isA<LoggerApiImpl>());
   });
 }