+++ packages/fluent_environment/fluent_environment/CHANGELOG.md
@@ -35,3 +35,7 @@
 ## 0.2.0
 
 * Upgrade package to flutter version v3.35
+
+## 0.3.0
+
+* Upgrade fluent_sdk version to v0.4.0